remove iterator traits header
