+++ core/include/detray/utils/ranges/subrange.hpp
@@ -21,31 +21,32 @@ namespace detray::ranges {
 ///
 /// @tparam range_t the iterable which to constrain to a subrange.
 template <typename range_t>
-struct subrange_view : public ranges::view_interface<subrange_view<range_t>> {
+struct subrange : public ranges::view_interface<subrange<range_t>> {
 
     using iterator_t = typename detray::ranges::iterator_t<range_t>;
     using const_iterator_t = typename detray::ranges::const_iterator_t<range_t>;
     using range_size_t = typename detray::ranges::range_size_t<range_t>;
 
     /// Default constructor
-    constexpr subrange_view() = default;
+    constexpr subrange() = default;
 
     /// Construct from a range: The subrange spans the entire range
     ///
     /// @param range container to iterate over
     template <typename deduced_range_t>
-    DETRAY_HOST_DEVICE constexpr explicit subrange_view(deduced_range_t &&range)
-        : m_start{detray::ranges::begin(range)},
-          m_end{detray::ranges::end(range)} {}
+    DETRAY_HOST_DEVICE constexpr explicit subrange(deduced_range_t &&range)
+        : m_start{detray::ranges::begin(std::forward<deduced_range_t>(range))},
+          m_end{detray::ranges::end(std::forward<deduced_range_t>(range))} {}
 
     /// Construct from a range and a single position.
     ///
     /// @param range container to iterate over
     /// @param pos start and end position for iteration
     template <typename deduced_range_t>
-    DETRAY_HOST_DEVICE constexpr subrange_view(deduced_range_t &&range,
-                                               range_size_t pos)
-        : m_start{detray::ranges::begin(range) + pos},
+    DETRAY_HOST_DEVICE constexpr subrange(deduced_range_t &&range,
+                                          range_size_t pos)
+        : m_start{detray::ranges::begin(std::forward<deduced_range_t>(range)) +
+                  pos},
           m_end{detray::ranges::next(m_start)} {}
 
     /// Construct from a range and start/end positions
@@ -53,10 +54,12 @@ struct subrange_view : public ranges::view_interface<subrange_view<range_t>> {
     /// @param range container to iterate over
     /// @param pos start and end position for iteration
     template <typename deduced_range_t, typename index_range_t>
-    DETRAY_HOST_DEVICE constexpr subrange_view(deduced_range_t &&range,
-                                               index_range_t &&pos)
-        : m_start{detray::ranges::begin(range) + detray::detail::get<0>(pos)},
-          m_end{detray::ranges::begin(range) + detray::detail::get<1>(pos)} {}
+    DETRAY_HOST_DEVICE constexpr subrange(deduced_range_t &&range,
+                                          index_range_t &&pos)
+        : m_start{detray::ranges::begin(std::forward<deduced_range_t>(range)) +
+                  detray::detail::get<0>(pos)},
+          m_end{detray::ranges::begin(std::forward<deduced_range_t>(range)) +
+                detray::detail::get<1>(pos)} {}
 
     /// Construct from a range and start/end positions
     ///
@@ -65,11 +68,12 @@ struct subrange_view : public ranges::view_interface<subrange_view<range_t>> {
     template <typename deduced_range_t, typename volume_t,
               std::enable_if_t<std::is_same_v<volume_t::volume_def, volume_t>,
                                bool> = true>
-    DETRAY_HOST_DEVICE subrange_view(deduced_range_t &&range, volume_t &&vol) {
+    DETRAY_HOST_DEVICE subrange(deduced_range_t &&range, volume_t &&vol) {
         dindex_range r = vol.template range<
             typename detray::ranges::range_value_t<deduced_range_t>>();
 
-        auto start = detray::ranges::begin(range);
+        auto start =
+            detray::ranges::begin(std::forward<deduced_range_t>(range));
 
         m_start = start + detray::detail::get<0>(r);
         m_end = start + detray::detail::get<1>(r);
@@ -79,10 +83,9 @@ struct subrange_view : public ranges::view_interface<subrange_view<range_t>> {
     ///
     /// @param start container to iterate over
     /// @param end start and end position for iteration
-    DETRAY_HOST_DEVICE constexpr subrange_view(
-        typename detray::ranges::iterator_t<range_t> &&start,
-        typename detray::ranges::iterator_t<range_t> &&end)
-        : m_start{start}, m_end{end} {}
+    DETRAY_HOST_DEVICE constexpr subrange(iterator_t &&start, iterator_t &&end)
+        : m_start{std::forward<iterator_t>(start)},
+          m_end{std::forward<iterator_t>(end)} {}
 
     /// @return start position of range.
     DETRAY_HOST_DEVICE
@@ -99,53 +102,6 @@ struct subrange_view : public ranges::view_interface<subrange_view<range_t>> {
     /// Start and end position of the subrange
     iterator_t m_start, m_end;
 };
-
-namespace views {
-
-/// @brief interface type to construct a @c subrange_view with CTAD
-template <typename range_t>
-struct subrange : public ranges::subrange_view<range_t> {
-
-    using base_type = ranges::subrange_view<range_t>;
-    using iterator_t = typename base_type::iterator_t;
-    using const_iterator_t = typename base_type::const_iterator_t;
-    using range_size_t = typename base_type::range_size_t;
-
-    constexpr subrange() = default;
-
-    template <typename deduced_range_t>
-    DETRAY_HOST_DEVICE constexpr explicit subrange(deduced_range_t &&range)
-        : base_type(std::forward<deduced_range_t>(range)) {}
-
-    template <typename deduced_range_t>
-    DETRAY_HOST_DEVICE constexpr subrange(deduced_range_t &&range,
-                                          range_size_t pos)
-        : base_type(std::forward<deduced_range_t>(range), pos) {}
-
-    template <typename deduced_range_t, typename index_range_t>
-    DETRAY_HOST_DEVICE constexpr subrange(deduced_range_t &&range,
-                                          index_range_t &&pos)
-        : base_type(std::forward<deduced_range_t>(range),
-                    std::forward<index_range_t>(pos)) {}
-
-    template <typename deduced_range_t, typename volume_t,
-              std::enable_if_t<std::is_same_v<volume_t::volume_def, volume_t>,
-                               bool> = true>
-    DETRAY_HOST_DEVICE constexpr subrange(deduced_range_t &&range,
-                                          volume_t &&vol)
-        : base_type(std::forward<deduced_range_t>(range),
-                    std::forward<volume_t>(vol)) {}
-
-    DETRAY_HOST_DEVICE constexpr subrange(
-        typename detray::ranges::iterator_t<range_t> &&start,
-        typename detray::ranges::iterator_t<range_t> &&end)
-        : base_type(
-              std::forward<typename detray::ranges::iterator_t<range_t>>(start),
-              std::forward<typename detray::ranges::iterator_t<range_t>>(end)) {
-    }
-};
-
-// deduction guides
 
 template <typename deduced_range_t>
 DETRAY_HOST_DEVICE subrange(deduced_range_t &&range)->subrange<deduced_range_t>;
@@ -171,7 +127,5 @@ DETRAY_HOST_DEVICE subrange(
     typename detray::ranges::iterator_t<deduced_range_t> &&start,
     typename detray::ranges::iterator_t<deduced_range_t> &&end)
     ->subrange<deduced_range_t>;
-
-}  // namespace views
 
 }  // namespace detray::ranges
